Creating rectangle on red pixels works - needs to be calibrated

+++ SigmaDetec/Point.cs
@@ -8,9 +8,9 @@ namespace SigamDetec
 {
     public class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public Point(int x, int y)
+        public double X { get; set; }
+        public double Y { get; set; }
+        public Point(double x, double y)
         {
             try
             {
